requirement.txt
aiodns==3.0.0
aiohttp==3.8.5
async-timeout==4.0.3
attrs==23.1.0
cffi==1.15.1
chardet==5.2.0
cryptography==41.0.3
discord.py==2.3.2
idna==3.4
multidict==6.0.4
pycparser==2.21
yarl==1.9.2
